Removed PWM Mod 1/2
The main reason I removed it to anyone who cares is that it doesn't support A16 yet, and I don't really see it getting support in the foreseeable future. Although, I may just be acting pessimistic, as Sultan updated an hour ago, so until support is added, this is just a personal use that has a more updated SUSFS version.